% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{hydrodata_list}
\alias{hydrodata_list}
\title{List available datasets}
\usage{
hydrodata_list(year = NULL)
}
\arguments{
\item{year}{Optional year filter (character or numeric).}
}
\value{
A data frame of available datasets.
}
\description{
List available datasets
}
